backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
openai==1.3.7
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-json-logger==2.0.7
pillow==10.1.0
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0